requirements.txt
ollama==0.1.0
streamlit==1.23.0
pandas==1.5.3
PyPDF2==1.28.0
python-docx==0.8.11
openpyxl==3.0.10
